Change the format of the summary shown in the PRs
Previously the format of the summary would change depending of the state
of the build jobs (building, failed, succeeded). This means that a "3 / 4
building" could at the end of the test run turn into "1 / 4 failed".
Which can be confusing.

With this patch the summary always shows the number of finished build jobs
and the overall job count. In case of build failures a count of failed
builds is shown as well.

The indicator for the overall state of a test run is already covered by
the GitHub status-icon.