Make openQA trigger script stateless
We always look at the last build result on OBS. Instead of comparing filenames to
the last build we have seen previously (state), we look if the build is older
than the frequency we check. If yes, it's a build we haven't seen.

Also: Ruby'fy the script and configure via environment